Implement settings update

+++ routes/web.php
@@ -15,36 +15,36 @@ Route::middleware(['auth', 'verified'])->prefix('admin')->group(function () {
     Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
 
     // Users Management
-    Route::get('/users', [UserController::class, 'index'])->name('users.index');
-    Route::get('/users/create', [UserController::class, 'create'])->name('users.create');
-    Route::post('/users', [UserController::class, 'store'])->name('users.store');
-    Route::get('/users/{id}', [UserController::class, 'show'])->name('users.show');
-    Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('users.edit');
-    Route::put('/users/{id}', [UserController::class, 'update'])->name('users.update');
-    Route::delete('/users/{id}', [UserController::class, 'delete'])->name('users.delete');
+    Route::get('/users', [UserController::class, 'index'])->name('admin.users.index');
+    Route::get('/users/create', [UserController::class, 'create'])->name('admin.users.create');
+    Route::post('/users', [UserController::class, 'store'])->name('admin.users.store');
+    Route::get('/users/{id}', [UserController::class, 'show'])->name('admin.users.show');
+    Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('admin.users.edit');
+    Route::put('/users/{id}', [UserController::class, 'update'])->name('admin.users.update');
+    Route::delete('/users/{id}', [UserController::class, 'delete'])->name('admin.users.delete');
 
     // Update User's password
-    Route::get('/users/{id}/password', [UserController::class, 'editPassword'])->name('users.editPassword');
-    Route::put('/users/{id}/password', [UserController::class, 'updatePassword'])->name('users.updatePassword');
+    Route::get('/users/{id}/password', [UserController::class, 'editPassword'])->name('admin.users.editPassword');
+    Route::put('/users/{id}/password', [UserController::class, 'updatePassword'])->name('admin.users.updatePassword');
 
     // Cars
-    Route::get('/cars', [CarController::class, 'index'])->name('cars.index');
-    Route::get('/cars/create', [CarController::class, 'create'])->name('cars.create');
-    Route::post('/cars', [CarController::class, 'store'])->name('cars.store');
-    Route::get('/cars/{id}', [CarController::class, 'show'])->name('cars.show');
-    Route::get('/cars/{id}/edit', [CarController::class, 'edit'])->name('cars.edit');
-    Route::put('/cars/{id}', [CarController::class, 'update'])->name('cars.update');
-    Route::delete('/cars/{id}', [CarController::class, 'delete'])->name('cars.delete');
+    Route::get('/cars', [CarController::class, 'index'])->name('admin.cars.index');
+    Route::get('/cars/create', [CarController::class, 'create'])->name('admin.cars.create');
+    Route::post('/cars', [CarController::class, 'store'])->name('admin.cars.store');
+    Route::get('/cars/{id}', [CarController::class, 'show'])->name('admin.cars.show');
+    Route::get('/cars/{id}/edit', [CarController::class, 'edit'])->name('admin.cars.edit');
+    Route::put('/cars/{id}', [CarController::class, 'update'])->name('admin.cars.update');
+    Route::delete('/cars/{id}', [CarController::class, 'delete'])->name('admin.cars.delete');
 
     // Settings
-    Route::get('/settings', [SettingsController::class, 'index'])->name('settings.index');
-    Route::get('/settings/edit', [SettingsController::class, 'create'])->name('settings.edit');
-    Route::put('/settings', [SettingsController::class, 'update'])->name('settings.update');
+    Route::get('/settings', [SettingsController::class, 'index'])->name('admin.settings.index');
+    Route::get('/settings/edit', [SettingsController::class, 'edit'])->name('admin.settings.edit');
+    Route::put('/settings', [SettingsController::class, 'update'])->name('admin.settings.update');
 
     // Profile
-    Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
-    Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
-    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+    Route::get('/profile', [ProfileController::class, 'edit'])->name('admin.profile.edit');
+    Route::patch('/profile', [ProfileController::class, 'update'])->name('admin.profile.update');
+    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('admin.profile.destroy');
 });
 
 Route::get('/', [HomeController::class, 'index']);
